Updated education controller and route

+++ controllers/education_controller.js
@@ -1,3 +1,5 @@
+
+import { userModel } from "../models/user_model.js";
 import { educationModel } from "../models/education_model.js";
 import { educationSchema } from "../schema/education_schema.js";
 
@@ -22,22 +24,22 @@ export const postEducation = async (req, res) => {
 
 
 
-        const user = await User.findById(userSessionId);
+        const user = await user.findById(userSessionId);
         // Find the user in the database using the user ID from the session
         if (!user) {
             return res.status(404).send('User not found');
         }
 
         const education = await educationModel.create({ ...value, user: userSessionId });
-        
+
         user.education.push(education._id);
 
         // Add the ID of the newly created education document to the user's education array
         await user.save();
-        
+
 
         res.status(201).json({ education });
-        
+
 
 
     } catch (error) {
@@ -73,7 +75,7 @@ export const getEducation = async (req, res) => {
 export const getSingleEducation = async (req, res) => {
     try {
         const getSingleEducation = await educationModel.findById(req.params.id, req.body, { new: true });
-        
+
         res.status(200).json(getSingleEducation);
 
     } catch (error) {
@@ -83,20 +85,32 @@ export const getSingleEducation = async (req, res) => {
 
 
 // Endpoint to update the details of an education
-export const updateEducation = async (req, res, next) => {
+export const updateEducation = async (req, res) => {
     try {
 
-        const { error } = educationSchema.validate(req.body);
+        const { error, value } = educationSchema.validate(req.body);
         if (error) {
             return res.status(400).send(error.details[0].message);
         }
 
-        const updateEducation = await educationModel.findByIdAndUpdate(req.params.id, req.body, { new: true });
-        res.status(200).json(updateEducation);
+        const userSessionId = req.session.user.id;
+
+        const user = await userModel.findById(userSessionId);
+        if (!user) {
+            return res.status(404).send("User not found");
+        }
+
+        const Education = await Education.findByIdAndUpdate(req.params.id, value, { new: true });
+        if (!Education) {
+            return res.status(404).send("Education not found");
+        }
+
+        res.status(201).json({ Education });
+
 
 
     } catch (error) {
-        next(error);
+        return res.status(500).json({ error })
     }
 }
 
@@ -105,13 +119,24 @@ export const updateEducation = async (req, res, next) => {
 export const deleteEducation = async (req, res, next) => {
     try {
 
-        const deleteEducation = await educationModel.findByIdAndDelete(req.params.id);
-        if (!deleteEducation) {
-            return res.status(404).send('Education not found');
+
+        const userSessionId = req.session.user.id;
+        const user = await userModel.findById(userSessionId);
+        if (!user) {
+            return res.status(404).send("User not found");
         }
-        res.status(200).json('Education deleted successfully');
+
+        const education = await educationModel.findByIdAndDelete(req.params.id);
+        if (!education) {
+            return res.status(404).send("Education not found");
+        }
+
+        user.education.pull(req.params.id);
+        await user.save();
+        res.status(200).json("Education deleted ");
+
 
     } catch (error) {
-        next(error);
+        return res.status(500).json({ error })
     }
-}
+};
